Database Wrapper
The database wrapper hooks and interfaces with the sqlite3 dastabase. Takes in the file path to the current sqlite3 database.

getProjectById, updateProjectById, getAllProjects, getAllActiveProjects, getAllProjectsByStatus, getAllProjectsByCategory, getAllHiddenProjects All are used to gather project data from the project table.

getVolunteerLoginDetails, doesUsernameExist, doesEmailExist, createNewVolunteer, markVolunteerAsVerified all interact with the volunteers table, all do as the name explains.

saltAndHash is used to generate a salted and hashed password based on a provided password, while you can also overwrite the salt by passing a salt, this would be used when authenticating.

createNewVerificationCode generates and stores a email verification code in the verification_code table. Used for verifiing the email when emails are sent out, currently emails don't work and the link is just logged with the code. The code is salted and hashed and stored in the table, then the code is salted and hashed when sent by the user to validate it.

removeVerificationCode removes the verification code if it exits for the provided userId, for when you validate the email.