Project Tests
Adjusted the project component to check if it exists before updating
the content while also checks to see if the project id is correctly
passed before attempting to query the database for its existence.

Tests are only ran while its not being built on travis.

Tests cover exist, component updating and getContent.