Promise rejection
All promises now reject as if they where node errors, using new Error() as prefered, this is also now checked with the linter.

Adjusted the mocha tests to fix with the new style of database inputs.

Fixed mocha hanging with --exit (this is actually our problem for not cleaning up properly)

updated packages.