Late Emails
stored json file is now ignored as it is only relevent while deployed.

When verifying the connection to the email service, the solution will check for existing stored late emails that would be stored while someone triggers a email but the service is not active or online.

Default configuration file now contains the path to the stored json file of the late emails. We don't expect this file to ever contain content but in the event of the service going down we will use it to store attempts.