Email services
Administrators now have routes to be able to update the service client password and the service details for the emailing client without the need to restart the server.

This will also update the stored configuration file if the updates are processed and restarted without errors. This allows the server work again also once it restarts.

Note: if the requests are processed through with a 200 response you can expect all files to be updated otherwise with any thrown errors or non 200 reponses you can expect that the files and changes have not been processed.