Proning tests and connection
Connection is made on the creation of the class Database instead of by the code invoking this state, this is because we always will require the connection to be made, otherwise without it there is no need to create the volunteer class, resulting in duplicate code.

Proned unused tests used simply to buff code coverage.