Async await start
started the process to move over to a async and await process, there are a couple of linting issues to fix while also a single bug related to password reset codes.

Linting issues refer to .finally being used in the code for bluebird, as bluebird is no longer being used.